+++ src/models/pulpo.model.js
@@ -38,12 +38,20 @@ objSchema.pre('deleteOne', true, function () {
         yield estado_model_1.default.objModel.deleteMany({ refPulpo: objToDel._id });
     });
 });
+objSchema.pre('remove', function () {
+    return __awaiter(this, void 0, void 0, function* () {
+        const objToDel = this;
+        yield estado_model_1.default.objModel.deleteMany({ refPulpo: objToDel._id });
+    });
+});
 objSchema.pre('deleteOne', false, function () {
     return __awaiter(this, void 0, void 0, function* () {
         const filtro = this.getQuery();
         const modelo = this.model;
         const objToDel = yield modelo.findOne(filtro);
-        yield estado_model_1.default.objModel.deleteMany({ refPulpo: objToDel._id });
+        if (objToDel != null) {
+            yield estado_model_1.default.objModel.deleteMany({ refPulpo: objToDel._id });
+        }
     });
 });
 objSchema.pre('deleteMany', function () {
